requirements.txt
aiogram
sqlalchemy
aiosqlite
httpx
tenacity
apscheduler
pytz
timezonefinder
python-dotenv
os